# Greatly lessen GC pressure along hot path in heap segment cache reading