This change greatly reduces GC pressure in the cache along common read paths.

**Overview:** The old code to read data off a page was shared between GetDataForAddress (where we know how much data we want) and GetDataFromAddressUntil (where we don't). To accomplish this it indirected the actual reading of the data through an Action passed to it. Then each caller would create a lambda to pass in for the action that would accomplish their specific style of read.

**Problem:** This worked for what it intended, but perf profiling showed it had the (obvious but) unintended side-effect of creating lots of short-lived objects (the lambda + the Action wrapper for the lambda) that put a lot of pressure on the GC. A single call was not bad, millions of calls was very bad.

**Solution:** We trade off the cost of the Action in the common case (where we know how many bytes we want to read) for a call to a virtual function that knows how to read bytes off the entries backing storage mechanism (a byte[] or an IntPtr). The cost of the virtual call is very low compared to the cost of the GC(s) created by the old code.

The GetDataFromAddressUntil still uses the lambda/Action approach because it is FAR less frequently called and I didn't feel like making each class have to handle the whole complicated 'matching against a terminating sequence/dealing with cross-page reads/carryover/etc...' stuff.