@leculver 2020-06-18
Yep definitely a small but noticeable improvement here.  (Note that I'm mostly just testing for major regression here.  A change doesn't have to be a huge perf win to be merged...I'm only going to reject changes that are major regressions from the perf side.)

Before:

|                        Method |  CacheSize | UseOSMemoryFeatures |     Mean |    Error |   StdDev |
|------------------------------ |----------- |-------------------- |---------:|---------:|---------:|
|               HeapEnumeration |  134217728 |               False | 16.708 s | 0.2919 s | 0.1737 s |
| HeapEnumerationWithReferences |  134217728 |               False | 85.446 s | 1.4927 s | 0.3877 s |
|               HeapEnumeration | 1073741824 |               False | 11.740 s | 0.2210 s | 0.1462 s |
| HeapEnumerationWithReferences | 1073741824 |               False | 41.905 s | 0.8276 s | 0.4328 s |
|               HeapEnumeration | 4294967296 |               False |  7.864 s | 0.0864 s | 0.0224 s |
| HeapEnumerationWithReferences | 4294967296 |               False | 30.805 s | 0.5873 s | 0.1525 s |

|                              Method |  CacheSize | UseOSMemoryFeatures | Threads |       Mean |     Error |    StdDev |
|------------------------------------ |----------- |-------------------- |-------- |-----------:|----------:|----------:|
| ParallelEnumerateHeapWithReferences |  134217728 |               False |       1 | 6,117.5 ms | 240.61 ms | 159.15 ms |
| ParallelEnumerateHeapWithReferences |  134217728 |               False |       4 | 5,794.6 ms | 113.52 ms |  75.09 ms |
| ParallelEnumerateHeapWithReferences |  134217728 |               False |       8 | 5,860.4 ms | 110.33 ms |  72.97 ms |
| ParallelEnumerateHeapWithReferences | 1073741824 |               False |       1 |   945.4 ms |  17.42 ms |   6.21 ms |
| ParallelEnumerateHeapWithReferences | 1073741824 |               False |       4 |   842.8 ms |  15.43 ms |   4.01 ms |
| ParallelEnumerateHeapWithReferences | 1073741824 |               False |       8 |   817.9 ms |   9.05 ms |   2.35 ms |
| ParallelEnumerateHeapWithReferences | 4294967296 |               False |       1 |   964.1 ms |  54.81 ms |  32.62 ms |
| ParallelEnumerateHeapWithReferences | 4294967296 |               False |       4 |   829.5 ms |  12.96 ms |   3.37 ms |
| ParallelEnumerateHeapWithReferences | 4294967296 |               False |       8 |   817.5 ms |  14.24 ms |   3.70 ms |

After:

|                        Method |  CacheSize | UseOSMemoryFeatures |     Mean |    Error |   StdDev |
|------------------------------ |----------- |-------------------- |---------:|---------:|---------:|
|               HeapEnumeration |  134217728 |               False | 15.958 s | 0.4577 s | 0.3027 s |
| HeapEnumerationWithReferences |  134217728 |               False | 81.181 s | 1.4998 s | 0.7844 s |
|               HeapEnumeration | 1073741824 |               False | 11.346 s | 0.2817 s | 0.1863 s |
| HeapEnumerationWithReferences | 1073741824 |               False | 39.021 s | 0.6730 s | 0.2400 s |
|               HeapEnumeration | 4294967296 |               False |  7.666 s | 0.1979 s | 0.0514 s |
| HeapEnumerationWithReferences | 4294967296 |               False | 29.667 s | 0.5310 s | 0.2777 s |




|                              Method |  CacheSize | UseOSMemoryFeatures | Threads |       Mean |     Error |    StdDev |
|------------------------------------ |----------- |-------------------- |-------- |-----------:|----------:|----------:|
| ParallelEnumerateHeapWithReferences |  134217728 |               False |       1 | 6,107.9 ms | 182.18 ms | 120.50 ms |
| ParallelEnumerateHeapWithReferences |  134217728 |               False |       4 | 5,675.1 ms | 112.72 ms |  58.95 ms |
| ParallelEnumerateHeapWithReferences |  134217728 |               False |       8 | 5,841.9 ms | 112.57 ms |  66.99 ms |
| ParallelEnumerateHeapWithReferences | 1073741824 |               False |       1 |   969.7 ms |  23.76 ms |  15.72 ms |
| ParallelEnumerateHeapWithReferences | 1073741824 |               False |       4 |   805.8 ms |  36.79 ms |  19.24 ms |
| ParallelEnumerateHeapWithReferences | 1073741824 |               False |       8 |   799.0 ms |  48.18 ms |  28.67 ms |
| ParallelEnumerateHeapWithReferences | 4294967296 |               False |       1 |   958.3 ms |  18.12 ms |   9.48 ms |
| ParallelEnumerateHeapWithReferences | 4294967296 |               False |       4 |   797.2 ms |  11.79 ms |   4.21 ms |
| ParallelEnumerateHeapWithReferences | 4294967296 |               False |       8 |   791.5 ms |  14.31 ms |   6.35 ms |